relax date parsing
we should not fail if an invalid date is provided. it is actually
highly likely that there is some data in an statement that we think is
a date but is not depending on the bank.

if we can not parse the date we simply return the provided value - then
it is up to the user to make sense of the value.