unify amount handling
The statement line hasn't used the to_amount/to_amount_in_cents helper
yet.
Also we do only provide positive amount values. If needed the amount
can be multiplied with the sign (-1 for debit or 1 for a credit) to get
the value that changes the balance (positive for credits and negative
for debits)