MT940 can contain details about every single field - handle it accordingly
The previous implementation had an issue, as it only expected transactions to have details. The way it was implemented, it was adding every meta data found in field 86 to the last transaction it parsed. In our example, the end balance for each statement has some meta data. Due to the implementation, the meta data was assigned to the las transaction and therefore overriding its already assigned meta data.