Added strip_headers conifguration option to remove SWIFT header data before parsing
some MT940 files come with additional header information. Currently
parsing failes when these headers are present. This options tries to
clean up the file and remove the headers before parsing as mostly this
information is not needed.
By default this option is disabled and should only be enabled when these
header files are present.

As a further improvement we should support parsing of the header data
aswell.